+++ src/app/api/user/auth/route.ts
@@ -3,11 +3,15 @@ import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 import { cookies } from "next/headers"; 
 
-const { AIRTABLE_API_KEY, AIRTABLE_BASE_ID, AITABLE_TABLE_USERS, JWT_SECRET } = process.env;
-const AIRTABLE_API_URL = `https://api.airtable.com/v0/${AIRTABLE_BASE_ID}/${AITABLE_TABLE_USERS}`;
+const { AIRTABLE_API_KEY, AIRTABLE_BASE_ID, AIRTABLE_TABLE_USER, JWT_SECRET } = process.env;
+const AIRTABLE_API_URL = `https://api.airtable.com/v0/${AIRTABLE_BASE_ID}/${AIRTABLE_TABLE_USER}`;
 
 export async function POST(request: Request) {
   try {
+
+    if (!AIRTABLE_API_KEY || !AIRTABLE_BASE_ID || !AIRTABLE_TABLE_USER || !JWT_SECRET) {
+      return NextResponse.json({ error: "Missing environment variables" }, { status: 500 });
+    }
     const { email, password } = await request.json();
 
     if (!email || !password) {
@@ -57,40 +61,5 @@ export async function POST(request: Request) {
   }
 }
 
-export async function GET(request: Request) {
-    try {
-      const cookieStore = await cookies();
-      const token = cookieStore.get("token");
-  
-      if (!token) {
-        return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
-      }
-  
-      if (!JWT_SECRET) {
-        throw new Error("JWT_SECRET is not defined in environment variables");
-      }
-  
-      
-      jwt.verify(token.value, JWT_SECRET);
-  
-      
-      const res = await fetch(AIRTABLE_API_URL, {
-        headers: {
-          Authorization: `Bearer ${AIRTABLE_API_KEY}`,
-        },
-      });
-  
-      if (!res.ok) {
-        return NextResponse.json({ error: "Failed to fetch users from Airtable" }, { status: res.status });
-      }
-  
-      const data = await res.json();
-  
-      
-      return NextResponse.json({ users: data.records });
-    } catch (error) {
-      console.error(error);
-      return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
-    }
-  }
+
