+++ src/app/api/user/findAll/route.ts
@@ -1,0 +1,45 @@
+import { NextResponse } from "next/server";
+import jwt from "jsonwebtoken";
+import { cookies } from "next/headers";
+
+const { AIRTABLE_API_KEY, AIRTABLE_BASE_ID, AIRTABLE_TABLE_USER, JWT_SECRET } = process.env;
+
+const AIRTABLE_API_URL = `https://api.airtable.com/v0/${AIRTABLE_BASE_ID}/${AIRTABLE_TABLE_USER}`;
+
+export async function GET(request: Request) {
+  try {
+    if (!AIRTABLE_API_KEY || !AIRTABLE_BASE_ID || !AIRTABLE_TABLE_USER || !JWT_SECRET) {
+      return NextResponse.json({ error: "Environment variables missing" }, { status: 500 });
+    }
+
+    const cookieStore = await cookies();
+    const token = cookieStore.get("token");
+
+    if (!token) {
+      return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+    }
+
+    try {
+      jwt.verify(token.value, JWT_SECRET);
+    } catch {
+      return NextResponse.json({ error: "Invalid token" }, { status: 401 });
+    }
+
+    const res = await fetch(AIRTABLE_API_URL, {
+      headers: {
+        Authorization: `Bearer ${AIRTABLE_API_KEY}`,
+      },
+    });
+
+    if (!res.ok) {
+      return NextResponse.json({ error: "Failed to fetch users from Airtable" }, { status: res.status });
+    }
+
+    const data = await res.json();
+
+    return NextResponse.json({ users: data.records });
+  } catch (error) {
+    console.error(error);
+    return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
+  }
+}
